backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.116.1",
    "psycopg2-binary>=2.9.9",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0"
]
